+++ src/constants/contactsInfo.ts
@@ -1,0 +1,18 @@
+import MailIcon from '@/assets/icons/mail-icon.svg';
+import MarkerIcon from '@/assets/icons/marker-icon.svg';
+import PhoneIcon from '@/assets/icons/phone-icon.svg';
+
+export const CONTACTS_INFO = [
+  {
+    Icon: MarkerIcon,
+    desc: '8558 Green Rd., LA',
+  },
+  {
+    Icon: PhoneIcon,
+    desc: '+1 (603) 555-0123',
+  },
+  {
+    Icon: MailIcon,
+    desc: 'example@gmail.com',
+  },
+];
