--- src/pages/BankCard/index.ts
@@ -1,2 +1,0 @@
-export { default as BankCard } from './BankCard';
-export { default as BankCardLazy } from './BankCard.lazy';
