--- src/pages/Contacts/Contacts.lazy.tsx
@@ -1,13 +1,0 @@
-import { lazy, Suspense } from 'react';
-
-import { PageLoader } from '@/components/PageLoader';
-
-const Contacts = lazy(() => import('./Contacts'));
-
-export default function ContactsLazy() {
-  return (
-    <Suspense fallback={<PageLoader />}>
-      <Contacts />
-    </Suspense>
-  );
-}
