--- src/pages/Contacts/index.ts
@@ -1,2 +1,0 @@
-export { default as Contacts } from './Contacts';
-export { default as ContactsLazy } from './Contacts.lazy';
