--- src/pages/Home/index.ts
@@ -1,2 +1,0 @@
-export { default as Home } from './Home';
-export { default as HomeLazy } from './Home.lazy';
