--- src/pages/NotFound/index.ts
@@ -1,2 +1,0 @@
-export { default as NotFound } from './NotFound';
-export { default as NotFoundLazy } from './NotFound.lazy';
