--- src/pages/Timeline/Timeline.lazy.tsx
@@ -1,13 +1,0 @@
-import { lazy, Suspense } from 'react';
-
-import { PageLoader } from '@/components/PageLoader';
-
-const Timeline = lazy(() => import('./Timeline'));
-
-export default function TimelineLazy() {
-  return (
-    <Suspense fallback={<PageLoader />}>
-      <Timeline />
-    </Suspense>
-  );
-}
