+++ src/router/index.tsx
@@ -1,27 +1,31 @@
+import { Suspense } from 'react';
 import { Route, Routes } from 'react-router-dom';
 
-import { BankCardLazy } from '@/pages/BankCard';
-import { ContactsLazy } from '@/pages/Contacts';
-import { HomeLazy } from '@/pages/Home';
-import { NotFoundLazy } from '@/pages/NotFound';
-import { TimelineLazy } from '@/pages/Timeline';
+import { PageLoader } from '@/components/PageLoader';
+import { BankCard } from '@/pages/BankCard';
+import { Contacts } from '@/pages/Contacts';
+import { Home } from '@/pages/Home';
+import { NotFound } from '@/pages/NotFound';
+import { Timeline } from '@/pages/Timeline';
 
 import { AppRoutes } from './routes';
 
 const router = [
-  { path: AppRoutes.HOME, element: <HomeLazy /> },
-  { path: AppRoutes.TIMELINE, element: <TimelineLazy /> },
-  { path: AppRoutes.BANK_CARD, element: <BankCardLazy /> },
-  { path: AppRoutes.CONTACTS, element: <ContactsLazy /> },
-  { path: AppRoutes.NOT_FOUND, element: <NotFoundLazy /> },
+  { path: AppRoutes.HOME, element: <Home /> },
+  { path: AppRoutes.TIMELINE, element: <Timeline /> },
+  { path: AppRoutes.BANK_CARD, element: <BankCard /> },
+  { path: AppRoutes.CONTACTS, element: <Contacts /> },
+  { path: AppRoutes.NOT_FOUND, element: <NotFound /> },
 ];
 
 export const Router = () => {
   return (
-    <Routes>
-      {router.map(({ path, element }) => (
-        <Route key={path} path={path} element={element} />
-      ))}
-    </Routes>
+    <Suspense fallback={<PageLoader />}>
+      <Routes>
+        {router.map(({ path, element }) => (
+          <Route key={path} path={path} element={element} />
+        ))}
+      </Routes>
+    </Suspense>
   );
 };
